agentic_framework/requirements.txt
fastmcp>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
azure-identity>=1.15.0
azure-cosmos>=4.5.1
openai>=1.6.0
httpx>=0.25.0
structlog>=23.2.0
tenacity>=8.2.3
gremlin-python>=3.7.0
pyodbc>=5.0.1
rapidfuzz>=3.5.0
PyJWT[crypto]>=2.8.0
